requirements.txt
flask 
flask-sqlalchemy 
flask-login
paho-mqtt
pillow
numpy
opencv-python
pyodbc
